Issue #258  - Location field in the backend.
Now a StreamPost can include a location.  If it does, it will automatically update the user's last known location.  Also added a Location model to hold all of the locations.  Figured we can before the next cruise add all of the locations of the Cruise, then both Events and this can utilize the same data.

I tried to implement this in the front end but I totally do not understand Ember and any of its design decisions.  I just can't follow any of its logic.